+++ src/Structures/Proxy/ChannelSend.ts
@@ -1,0 +1,29 @@
+import { 
+    TextChannel,
+    Message,
+    DMChannel,
+    NewsChannel
+} from "discord.js";
+import { Logger } from "../Logger";
+
+const logger = new Logger('ChannelSend Proxy');
+
+TextChannel.prototype.send = new Proxy(TextChannel.prototype.send, {
+    async apply(target, thisArg: TextChannel | DMChannel, args) {
+        try {
+            const m: Message = await target.call(thisArg, ...args);
+            return m;
+        } catch(e) {
+            logger.log(`
+            Channel Type: "${thisArg.type}"
+            | Channel ID: "${thisArg.id}"
+            | User ID: "${thisArg instanceof DMChannel ? thisArg.recipient.id : 'Not DMs'}"
+            | Error: "${e.toString()}"
+            `.split(/\n\r|\n|\r/g).map(e => e.trim()).join(' ').trim());
+            return null;
+        }
+    }
+});
+
+DMChannel.prototype.send = TextChannel.prototype.send;
+NewsChannel.prototype.send = NewsChannel.prototype.send;